Removed the correction for the title with the quote
because I don't like the correction.

Optimized the code for the title and author, now just use a static getter for the title and author for the whole application.
Changed the design of the scrollBar and always show it, to not create a lag and show the scrollbar if you don't need to show the scrollBar (when you don't have a book for example)